refactor(registration): extract shared reserveClassRegistration helper
Behavior-preserving extraction of createRegistration's steps 1-5 (validate →
required-agreements check → price → atomic capacity-transaction reserving a
`pending` registration) into `@maple/firebase/functions`
reserveClassRegistration, so the forthcoming hosted-checkout callable reserves
a spot identically. Single-sourcing the capacity logic prevents the two
payment paths from diverging (divergence = overbooking).

- New registration-reservation.utility.ts (reserveClassRegistration,
  validateRequiredAgreements, generateConfirmationNumber, RegistrationReservationInput).
  Input type is Omit<CreateRegistrationRequest,'paymentNonce'> so both the card
  flow (has nonce) and hosted flow (no nonce) satisfy it.
- createRegistration now calls the helper and keeps only its payment +
  agreement-processing + email steps.

Verified: maple-square codebase typechecks clean; registration integration
suite green (60/60). No behavior change.
