+++ libs/firebase/functions/src/lib/index.ts
@@ -76,6 +76,14 @@ export {
 // Email utilities
 export { isE2ETestEmail } from './email.utility';
 
+// Shared class-registration reservation (card + hosted-checkout paths)
+export {
+  reserveClassRegistration,
+  validateRequiredAgreements,
+  generateConfirmationNumber,
+  type ReserveRegistrationResult,
+} from './registration-reservation.utility';
+
 // Per-family calendar subscription utilities
 export {
   FAMILY_CALENDAR_FEED_PATH_PREFIX,
